share manifest types
