+++ src/utils/fs.ts
@@ -1,0 +1,1 @@
+export type PathLike = string;
